requirements/required.txt
Django==4.2.3
djangorestframework==3.14.0
uvicorn==0.22.0
